requirements.txt
accelerate>=0.21.0
appdirs
loralib
bitsandbytes
black
black[jupyter]
datasets
fire
git+https://github.com/huggingface/peft.git
transformers>=4.28.0
sentencepiece
gradio
antlr4-python3-runtime
protobuf
nltk
